# spec/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/pico-units:
    post:
      description: Register a Pico unit by handle after verifying it is reachable via mDNS.
      operationId: PicoUnitsController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePicoUnitDto'
      responses:
        '201':
          description: Pico unit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PicoUnit'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '409':
          description: A Pico unit with the same handle already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '424':
          description: Pico unit not reachable via mDNS
          content:
            application/json:
              schema:
                example:
                  statusCode: 424
                  error: Failed Dependency
                  description: Pico unit not reachable via mDNS
                  timestamp: '2024-09-14T22:00:00.000Z'
                  path: /path/of/the/failing/route
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Manually add a Pico Unit
      tags:
        - pico-units
    get:
      description: Retrieve all the pico units stored in the database, with minimal filtering options
      operationId: PicoUnitsController_list_v1
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: enabled
          required: false
          in: query
          description: Filter by enabled
          schema:
            type: boolean
        - name: q
          required: false
          in: query
          description: Search in handle/name/ip
          schema:
            type: string
      responses:
        '200':
          description: List with pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PicoUnitListResponseDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: List Pico Units (paginated)
      tags:
        - pico-units
  /v1/pico-units/announce:
    post:
      description: Called by Pico units on boot with hardware/network metadata. Upserts by handle.
      operationId: PicoUnitsController_announce_v1
      parameters:
        - name: X-Pico-Secret
          in: header
          description: Shared secret for Pico hardware announcements
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncePicoUnitDto'
      responses:
        '201':
          description: Pico unit registered/updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PicoUnit'
        '401':
          description: Missing or invalid X-Pico-Secret header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '409':
          description: Existing unit cannot be updated without an IP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Announce a Pico Unit (hardware)
      tags:
        - pico-units
...
info:
  title: mushpi-server
  description: A backend for the Mushroom Pi project
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    secret:
      type: http
      scheme: bearer
      in: header
      name: authorization
  schemas:
...
    ErrorDto:
      type: object
      properties: {}
    CreatePicoUnitDto:
      type: object
      properties:
        handle:
          type: string
          example: unit-01
          description: Handle (mDNS name) of the Pico unit to add manually
          minLength: 1
          maxLength: 64
      required:
        - handle
    Readings:
      type: object
      properties: {}
    PicoUnit:
      type: object
      properties:
        host:
          type: string
          description: mDNS hostname derived from handle
          example: unit-01.local
        address:
          type: string
          description: Local mDNS address for the pico unit's REST API
          example: http://unit-01.local:5000
        ipAddress:
          type: string
          description: IP-based address for the pico unit (fallback)
          example: http://192.168.1.50:5000
          nullable: true
        latest_reading:
          nullable: true
          description: Latest reading row for this unit (computed, not stored)
          allOf:
            - $ref: '#/components/schemas/Readings'
      required:
        - host
        - address
        - ipAddress
        - latest_reading
    AnnouncePicoUnitDto:
      type: object
      properties:
        handle:
          type: string
          example: unit-01
          description: Name for the unit as it is stated in its own software. Should not be changed.
          minLength: 1
          maxLength: 64
        ip:
          type: string
          description: IPv4 address where the Pico API is reachable. Provided by the unit on announcement.
          example: 192.168.1.50
        port:
          type: number
          example: 5000
          minimum: 1
          maximum: 65535
          default: 5000
          description: Pico API port
        micropython_version:
          type: string
          minLength: 1
          maxLength: 255
          description: Version tag for the Micropython executed by the unit
          examples:
            - v1.26.0 on 2025-08-09 (GNU 14.2.0 MinSizeRel)
            - 1.26.0.
        software_version:
          type: string
          minLength: 1
          maxLength: 255
          description: Version tag for the custom software executed by the unit
          examples:
            - 0.1.0
        board:
          type: string
          minLength: 1
          maxLength: 255
          description: Name of the unit's board
          examples:
            - Raspberry Pi Pico 2 W with RP2350
        board_total_mem_byte:
          type: number
          description: Size of the unit's memory in bytes
        board_total_fs_byte:
          type: number
          description: Size of the unit's file system in bytes
        board_cpu_freq_mhz:
          type: number
          description: CPU Frequency in MHz
        mac:
          type: string
          description: Physical MAC address of the unit
          examples:
            - 2c:cf:67:be:95:30
      required:
        - handle
        - port
    PicoUnitListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PicoUnit'
        page:
          type: number
          example: 1
          minimum: 1
        limit:
          type: number
          example: 20
          minimum: 1
          maximum: 500
        total:
          type: number
          example: 42
          minimum: 0
        pages:
          type: number
          example: 3
          minimum: 0
      required:
        - items
        - page
        - limit
        - total
        - pages
...